rknn/requirements.txt
fastapi==0.104.1
numpy==1.24.4
opencv-python-headless==4.6.0.66
python-dotenv==1.0.0
uvicorn==0.24.0.post1
python-multipart==0.0.6
Pillow==9.5.0
rapidocr-onnxruntime
